pyproject.toml
[project]
name = "wi-parser"
version = "1.0.0"
description = "A small utility for parsing images from Wildlife Insights website"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "tqdm>=4.67.3",
]
